Remove NLVSHierarchy
Refinement information is now propagated correctly such that the normal
variational solve can use GMG solver options.
