+++ firedrake/mg/__init__.py
@@ -5,4 +5,3 @@ from .function import *          # noqa: F401
 from .functionspace import *     # noqa: F401
 from .ufl_utils import *         # noqa: F401
 from .interface import *         # noqa: F401
-from .solver_hierarchy import *  # noqa: F401
